temp

+++ internal/domain/media/view.go
@@ -37,7 +37,7 @@ func ToDetailViewFromCreated(media databasegen.CreateMediaRow) *DetailView {
 
 func ToDetailViewFromFindByIDs(media databasegen.FindMediasByIDsRow) *DetailView {
 	return &DetailView{
-		ID:        int64(media.ID),
+		ID:        media.ID,
 		MediaType: Type(media.MediaType),
 		URL:       media.Url,
 		CreatedAt: media.CreatedAt,
